+++ packages/react/src/components/AnimatedHeader/assets/index.js
@@ -1,0 +1,31 @@
+/**
+ * @license
+ *
+ * Copyright IBM Corp. 2025
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+import dataFabricAnimatedLight from './animated/header_1312x738_data_fabric_light_03.json';
+import dataFabricAnimatedDark from './animated/header_1312x738_data_fabric_dark_06.json';
+import dataFabricStaticLight from './static/illustration_data_fabric_light_04.svg';
+import dataFabricStaticDark from './static/illustration_data_fabric_dark_05.svg';
+import watsonXAnimatedLight from './animated/header_1312x738_watsonx_data_light_03.json';
+import watsonXAnimatedDark from './animated/header_1312x738_watsonx_data_dark_02.json';
+import watsonXStaticLight from './static/illustration_watson_x_data_light_03.svg';
+import watsonXStaticDark from './static/illustration_watson_x_data_dark_03.svg';
+import wxbiaAnimatedLight from './animated/header_illustration_container_wxbia_light_08.json';
+import wxbiaAnimatedDark from './animated/header_illustration_container_wxbia_dark_05.json';
+
+export {
+  dataFabricAnimatedLight,
+  dataFabricAnimatedDark,
+  dataFabricStaticLight,
+  dataFabricStaticDark,
+  watsonXAnimatedLight,
+  watsonXAnimatedDark,
+  watsonXStaticLight,
+  watsonXStaticDark,
+  wxbiaAnimatedLight,
+  wxbiaAnimatedDark,
+};